ugly: fix bad regression after CapConsole.ml -> Console.ml and ocaml-light
ocamldep in ocaml-light probably used to create an (incorrect) dep from Cap.cmo to
Console.cmo but that was still working because there was no cycle but
as soon as we merged CapConsol.ml in Console.ml, we now also add
Console.cmo depend on Cap.cmo hence the cycle.

The ugly trick is to help ocamldep by using different names (ugly).