lib_core: fix remaining linking issue because of changed stdlib func
undefined reference to `is_printable'
  removed see https://github.com/ocaml/ocaml/issues/6521
  https://github.com/ocaml/ocaml/commit/85030f805095e37e5cb87ac4223956030f0b5900
undefined reference to `hash_univ_param'
  new seeded_hash builtin instead, so easy fix
undefined reference to `caml_open_descriptor'
undefined reference to `channel_descriptor'

So mk and rc now compiles and links!
but segfault when I run it ... probably some external:
have a different types but it is not typechecked
or some missing annotations or whatever.

Still, at least it compiles and links! now maybe time
to go back to use ocaml-light for lib_core as it matches
more the actual stdlib in it.